+++ Server/controllers/EnrollmentController.js
@@ -1,4 +1,5 @@
 const EnrollmentModel = require('../models/Enrollment');
+const InstructorCourseModel = require("../models/InstructorCourse");
 
 class EnrollmentController {
     static async addStudentToCourse(req, res) {
@@ -24,9 +25,16 @@ class EnrollmentController {
     }
 
     static async addStudentToInstructorCourses(req, res) {
-        const { student_id, instructor_course_ids } = req.body;
-
         try {
+            const { student_id, instructor_course_ids } = req.body;
+
+            if (!student_id || instructor_course_ids.length <= 0)
+                return res.status(400).json({ success: false, message: 'All fields are required.' });
+
+            const isExist = await EnrollmentModel.checkStudentCourseExistence(student_id, instructor_course_ids);
+            if (isExist)
+                return res.status(409).json({ success: false, message: 'This instructor assigned to this course(s) already exists.' });
+
             const result = await EnrollmentModel.addStudentToInstructorCourses(student_id, instructor_course_ids);
             return res.json(result);
         } catch (err) {
@@ -36,9 +44,16 @@ class EnrollmentController {
     }
 
     static async deleteStudentEnrollment(req, res) {
-        const { id } = req.params;
-
         try {
+            const { id } = req.params;
+
+            if (!id)
+                return res.status(400).json({ success: false, message: 'ID is not define' });
+
+            const isExist = await EnrollmentModel.checkSingleStudentCourseExistence(id);
+            if (!isExist)
+                return res.status(409).json({ success: false, message: 'Student-Course relation is not exists.' });
+
             await EnrollmentModel.deleteStudentEnrollment(id);
             return res.json({ success: true });
         } catch (err) {
